Heeled Boots (Will Require 2da Adjustment/Merge)
These foot models are slightly adjusted from the boot set I added a few weeks ago. The heel has been elongated to give the shoes a natural feeling elevation for both genders, supporting a refined adventurer look. It will be occupying the 130 slot in the 2da for feet.

This will require line 130 to be adjusted in the foot appearance 2da which I've asked twohand to help me with as it is within a repository I do not have access to :)