Sashes 18 fix
I've done some optimization on the model, fixing vertices errors, and making adjustments to a majority of female pheno 1 models for better visual scaling and consistency. As as result of this work, the sizes on the models themselves have been slightly reduced by a single kilobyte or two.

A very slight contrast alteration has been made to the texture as well.

Belt Icons are also added.