January Clothing Drop (Requires Helm Script Update and 2da Update)
- 1 new set of pants.
- 1 new pelvis.
- 2 helmets.
- 2 fixed Kroatius helmets.
- 2 necks.

Since this is a large update I'll let the official changelog on discord speak for itself and keep the github entry clear and concise. This will require 3 2da updates, I will pass the 2da to Twohand or anyone else that is willing to get it updated. This will also require 2 new helmet numbers added to the script.