Fantasy Interiors Tileset - Path Node Improvement
This tileset has no path node work and 600 tiles. It'd be a huge effort to update all of them, so for now all that this PR does is updating the tiles used by some dungeons with (mostly) correct path nodes. If anything, it should fix their Line of Sight issues.